Controls
Added controls (don't work yet) to the right of the map. The controls
disable themself if they cannot be used. Will have to remember to add a
failsafe later, so that a user cannot use the controls even if they go
around this.
Also added a "How to play" section and a "symbol meaning" section.