Major development
Renamed all map instances in files to maze, since it is better named
that way.
Replaced "DatabaseException" with "FileDoesNotExistException"
Added code so that a maze can be saved and loaded from a file. The maze
is saved together with score, stepsleft, maxX maze tiles and maxY maze
tiles.
User is now recongnized thanks to a number in a cookie. The number is
also the name of the file for that users maze.
Some other minor changes as well.