MasterController, ScoreKeeper, Working Move
Major code update.
Added a master controller which meant changing a lot of code to work
with it (and to make stuff better).
Added a score keeper that tracks steps left and score
Added code to make the move controls work. You can even move to the next
maze! This still needs some work though, but overall it works. It won't
count extra steps for hazards as of now. Need to remake MazeHazards.php
for that to work.