Major development
Some major development:
- Removed unnecessary custom exceptions
- Changed how hazards are handled. It's now easy to add new hazard
classes by using the iMazeHazard interface. They are much easier to use
in code too (and better). Reworked code to work with this new change
- Added messages that can be displayed when an exception is thrown
- Added a step system, you now lose steps when going around the maze.
Hazards takes extra steps
- Added a simple message when you run out of steps that tells you that
the game is over. Also disables all buttons except the restart game
button.
- Rewrote portions of the code to make it easier to read, take less
space and be better
- Line of sight is now in all directions until the view hits a wall
- Added new CSS for the buttons. Not pretty, but it'll do the job.

I have some minor things left to do but I feel that I'm almost done with
this project. I'm going to rewrite the save system though, so that it
uses a database and is more secure (no login though, not that secure).
I'm also going to optimize and improve the code quality a bit.