Databaseexception
Added a new exception: DatabaseException. This exception is thrown
whenever there's something wrong when handling the database (for now, it
only checks if the connection succeded).
Also added some minor CSS and fixed so that the code actually removed
rows from the database.

Website is now uploaded to my server. The connection in this uploaded
DAL does not reflect the the passwords used on the server, but uses
standard settings instead (root/localhost).